requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
selenium==4.15.2
webdriver-manager==4.0.2
chromedriver-py==140.0.7339.207
google-cloud-vision==3.4.4
google-auth==2.23.4
google-api-python-client==2.108.0
reportlab==4.0.7
pydantic==2.5.0
httpx==0.25.2
pillow==10.1.0
python-dotenv==1.0.0
openai==1.3.0